evaluable modificar/borrar

+++ BD/Clase/Modificar y borrar/V2/accion_MoB.php
@@ -3,7 +3,19 @@ mysqli_report(MYSQLI_REPORT_ERROR);
 require "usargestiona.php";
 $codigo = $_REQUEST["codigo"];
 
-$consulta = "DELETE FROM productos WHERE codigoproducto='$codigo';";
+$accion = $_REQUEST["accion"];
+if ($accion == "Modificar") {
+  $descripcion = $_REQUEST["descripcion"];
+  $proveedor = $_REQUEST["proveedor"];
+  $preciocompra = $_REQUEST["preciocompra"];
+  $precioventa = $_REQUEST["precioventa"];
+  $stock = $_REQUEST["stock"];
+
+  $consulta = "UPDATE productos SET descripcionproducto='$descripcion',codigoproveedorproducto='$proveedor',
+  preciocompraproducto='$preciocompra', precioventaproducto='$precioventa', stockproducto='$stock' where codigoproducto='$codigo';";
+} else {
+  $consulta = "DELETE FROM productos WHERE codigoproducto='$codigo';";
+}
 
 if (!$resultado = $mysqli->query($consulta)) {
   echo "Lo sentimos. App falla<br>";
@@ -11,4 +23,4 @@ if (!$resultado = $mysqli->query($consulta)) {
   echo "Num.error: " . $mysqli->errno . "<br>";
   echo "Error: " . $mysqli->error . "<br>";
   exit;
-} else echo "<br>REGISTRO ELIMINADO<br>";
+} else echo "<br>ACCION REALIZADA CON EXITO ($accion).<br>";
